requirements.txt
python-telegram-bot==20.3
flask
requests
beautifulsoup4